Add coin and bronze brick

+++ MarioGame/Platform.h
@@ -8,7 +8,7 @@ class CPlatform : public CGameObject
 {
 public:
 
-	CPlatform(int widthUnit, int heightUnit);
+	CPlatform(float widthUnit, float heightUnit);
 	virtual void Render();
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
 };
